Lista de meus anúncios

+++ src/main/java/br/edu/ifrs/canoas/webapp/forms/ManagerSidebarItem.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ManagerSidebarItem {
     private String href;
-    private int size = 0;
+    private Long size = 0l;
     private boolean active = false;
 
     public ManagerSidebarItem(String href) {
